TcpClient bug (MQTT,HTTP client etc.)
Messages longer than 256 bytes were sent partly. (low value from int len only)
Tested on MQTT
(before work only those messages whose size: ( 6+ topic + payload) were less than 256 bytes)

+++ Sming/SmingCore/Network/TcpClient.cpp
@@ -74,7 +74,7 @@ bool TcpClient::sendString(String data, bool forceCloseAfterSent /* = false*/)
 	return send(data.c_str(), data.length(), forceCloseAfterSent);
 }
 
-bool TcpClient::send(const char* data, uint8_t len, bool forceCloseAfterSent /* = false*/)
+bool TcpClient::send(const char* data, uint16_t len, bool forceCloseAfterSent /* = false*/)
 {
 	if (state != eTCS_Connecting && state != eTCS_Connected) return false;
